feat: Attach labels to their input elements
This makes the labels more useful - they now label their actual input elements, and become clickable. This is in particular useful for the checkboxes, as you can now get away with simply clicking the text without having to hit the box.

The small CSS change makes sure the select pulldowns are still kept on their own line (since they're inline by default). The old component had a "hack" where the select element was wrapped in a div to achieve this.

Another option would be to generate a unique id inside each component and attach that id to the input element, and then use the `for` attribute on the label to attach it to the element. That would require getting a decent, unique id for each component, which would probably make another dependency necessary. Since a label can keep the element it is labeling inside itself instead and avoid adding extra complexity, this solution was chosen.