makes majority of the entrylist fields optional
valid entrylist entries can be quite small, e.g.

```
     {
       "drivers": [
         {
           "playerID": "...."
         }
       ],
       "isServerAdmin": 1
     }
```

the above valid entrylist entry is from a simgrid entrylist where all admins are added to the entrylist
even though they've not signed up for the race.

currently if these entrylists were imported as-is either via the GUI or API the result would end up as:

```
    {
      "drivers": [
        {
          "firstName": "",
          "lastName": "",
          "shortName": "",
          "driverCategory": 0,
          "playerID": "...",
          "nationality": 0
        }
      ],
      "raceNumber": 0,
      "forcedCarModel": 0,
      "overrideDriverInfo": 0,
      "isServerAdmin": 1,
      "customCar": "",
      "overrideCarModelForCustomCar": 0,
      "ballastKg": 0,
      "restrictor": 0,
      "defaultGridPosition": 0
    }
```

The main obvious problem here is '0' for the ForcedCarModel is a porsche car, whilst in reality the entrylist
from simgrid isn't forcing a car model, similar issues exist for some of the other settings which are not
defaulting to `-1` and instead defaulting to `0`.

The ideal scenario is for the entrylist to be exactly the same once imported via accweb, however due to the
checkboxes this is not really possible (checkboxes are either unchecked or checked via the GUI).

You could import via the API and get:

```
     {
       "drivers": [
         {
           "playerID": "...."
         }
       ],
       "isServerAdmin": 1
     }
```

However if you import via the GUI you will get:

```
    {
      "drivers": [
        {
          "playerID": "..."
        }
      ],
      "overrideDriverInfo": 0,
      "isServerAdmin": 1,
      "overrideCarModelForCustomCar": 0
    }
```

To be consistent I've instead opted to not set `omitempty` on `OverrideDriverInfo` and `OverrideCarModelForCustomCar`
which will keep the API and GUI consistent with each other to make future debugging/understanding easier. I don't
foresee any issues with importing a simgrid entrylist and having `overrideDriverInfo` and `overrideCarModelForCustomCar`
added by accweb unless ACC defaults in the future changes (very unlikely).